Add LICENSE file
🎉 Big News! Our ZATCA App is now under the MIT License! 🎉

To empower our amazing community and encourage the widest possible adoption, we have officially switched the license for this project from GPLv3 to the more permissive (MIT License).

What This Means for You

✅ Total Freedom: You can now integrate, customize, and use this app in any project, commercial or private, with fewer restrictions.

🚀 More Collaboration: This change makes it even easier for developers and businesses across Saudi Arabia to contribute, innovate, and build upon our shared work.

💡Wider Accessibility: Our goal is to make ZATCA compliance simple for everyone, and this license helps us achieve that.

Our mission remains the same: to provide the best open-source ZATCA integration for ERPNext. We're incredibly excited to see the amazing solutions you will build with this new freedom!

A huge thank you to everyone who has contributed so far. Let's continue working together to simplify ERP compliance in the Kingdom!